LCVW 88 finished
