+++ app/constants.py
@@ -1,3 +1,4 @@
 DATABASE_URL = 'localhost'
-MYSQL_USER = 'YOUR USER'
-MYSQL_PASS = 'YOUR PASSWORD'
+MYSQL_USER = 'cosa'
+MYSQL_PASS = 'lacosa'
+DATABASE_NAME = 'ing_2023'
