+++ app/tests/test_model_base.py
@@ -1,5 +1,5 @@
 import unittest
-
+from constants import DATABASE_NAME
 from model_base import ModelBase, initialize_database
 from create_mysql_db import create_database
 from player.models import Player
@@ -10,7 +10,7 @@ class TestModelBase(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         # Create and initialize the database
-        create_database('ing_2023')
+        create_database(DATABASE_NAME)
         initialize_database()
 
     def setUp(self):
